all: Actually print OS names in Titlecase
When printing the names of operating systems, we used the
strings.ToTitle function to turn the string into Titlecase - but
turns out that's not what strings.ToTitle does, instead it replaces all
Unicode letters with their title case (so for ASCII strings it's equal
to strings.ToUpper). What we want is the functionality of strings.Title,
but that's deprecated and golang.org/x/text/cases should be used, so
that's what we use now.