BREAKING: all: Let `--recover-ubsan` default to `true`
Since the previous commits ensure that we get inputs associated with
non-fatal UBSan findings, there is no longer a reason to deviate from
UBSan's default behavior by making findings fatal.

This breaking change is deemed to have minimal impact as it can only
increase the length of a fuzzing run.