build: Emit inputs on non-fatal UBSan findings
We hook `__sanitizer_report_error_summary` to run code whenever any
sanitizer produces a finding and call the callback provided to
`__sanitizer_set_death_callback` by libFuzzer to dump the current input
if the finding is a UBSan finding.

This currently applies only to the "cmake" and "other" build system.