fileutil: Remove CanonicalPath and its usages
Resolving the path of the project directory and all files and
directories in is causing problems, for example LCOV reports contain
unresolved source file paths, which we can't make relative to the
resolved project directory.

We found that the correct solution is to not resolve symlinks in paths.
The reason we did that in the first place was to make files and
directories from the host available in the minijail sandbox, which
should not be a problem anymore since we now make the root filesystem
of the host available as a read-only mount in the sandbox.

Removing this function requires adding calls to `filepath.Abs` in some
places in the `other` build system's build package.