finding: Copy input file instead of moving it
To avoid errors when two findings are produced for the same crashing
input, we now copy the input instead of moving it. No further clean up
should be necessary, because we already set `-artifact_prefix` to a
temporary directory (causing libFuzzer to place input files there)
which we clean up after the run.