pkg/dependencies: Remove unnecessary code
Setting CMakeDeps, MavenDeps and GradleDeps didn't have any effect. The
only thing that made CreateTestDeps work is that it overwrote the
Installed and GetVersion methods of the `all` struct.

This commit changes the comments to make the behavior clear and removes
unnecessary code.

Resetting the methods of the `all` struct in the TestMain is not
necessary, because each TestMain runs in its own process and any globals
changed in the package's tests doesn't change globals of other package's
tests.