integration-tests/bazel: Don't run tests in parallel
This assures that the tests are run in sequential order, making it
easier to avoid the tests interfering with each other.

The tests were blocking on other bazel commands to complete anyway, so
they could not actually run in parallel.

This also avoids that single tests time out because they are blocking on
bazel commands from other tests to complete.