cmd/remote-run: Make --bundle flag not configurable via viper
It wasn't configurable via environment variables before, because we
didn't call viper.BindPFlag. It's not useful to be configurable via the
config file because the path of the bundle created by `cifuzz bundle` is
not fixed.

Making this flag not configurable via viper also makes it easier to tell
the user that flags which only effect the build of the bundle are
ignored when the --bundle flag is used.