cmd/init: Remove leading newlines from instructions
Since commit 6bf67293752e99cd19911be52af4cd2307906804 we print the
instructions before trying to create the config file, so the
instructions are the first that's printed by the init command and should
therefore not start with a leading newline. That commit only removed the
newline from the bazel instructions though, so with this commit we also
remove the ones from the other instructions.