refactor: Move tools/cmake/cifuzz/{include,src}
These directories are not specific to CMake, so they should not live
below a "cmake" directory.

The src directory is split and moved to tools/{dumper,launcher,replayer}.

In preparation for installing the include directory to the top level of the
installation directory, we also put it in the top level of our project
directory, analogous to the share directory in our project root.

The approach of putting directories which are installed into the root of
the installation directory also in the root of the project directory
does not seem very common but also not particularly unusal. Examples of
other projects which do that to some extent are gh, which has a share
directory in the project root, and grpc, which has beside a include and
src directory (as usual for C/C++ projects) also an etc directory.