runner: Log the environment variables set by us
When using the sandbox, we used to set environment variables via the
process_wrapper, so they were printed in the verbose log output
(because there we print the commands we execute).

We now pass all environment variables from the host to the sandbox and
don't log them to avoid leaking secrets. To make debugging easier, this
commit again prints the environment variables set by us.