cmd/login: Don't support logging in via environment variable
The environment variable can be used alternatively to the access tokens
file. The login command adds tokens to the access tokens file but it
should not add a token which was specified via the environment variable.
Instead, we expect the environment variable to be set when commands
which use the token (like `cifuzz remote-run`) are executed.