executil: Always terminate child process group on terminating signals
When using executil.Command instead of exec.Command, the child process
is created in its own process group. This has the effect that it's not
terminated when the parent receives a terminating signal (e.g. SIGINT).
We handled that in some cases by registering a signal handler which
terminates the process group but it was too easy to forget that which
lead to processes not being terminated in tests. We now register a
signal handler in executil.Command.Start which terminates the process
group.